pyproject.toml
[project]
name = "agenix"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "flask-login>=0.6.3",
    "flask-sqlalchemy>=3.1.1",
    "langchain-community>=0.3.29",
    "langchain-experimental>=0.3.4",
    "langchain-google-genai>=2.1.10",
    "langchain-groq>=0.3.7",
    "langchain-postgres>=0.0.15",
    "langgraph>=0.6.6",
    "langgraph-checkpoint-postgres>=2.0.23",
    "psycopg-binary>=3.2.9",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
]
